+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 - always display html emails in black text on white background by default, fixes white text on white background when the OS was in dark theme mode
+- fix: webxdc loading progress bar finishing at 50% (not 100)
 
 <a id="1_36_2"></a>